refactor: extract group and chat logic into reusable components
