+++ front/src/components/Grupos/index.tsx
@@ -1,0 +1,58 @@
+export default function Grupos() {
+  return (
+    <div className="w-1/5 min-w-[180px] max-w-xs border-r border-blue-800 py-4 bg-[#0F172A] flex-shrink-0 flex flex-col">
+      <h2 className="font-bold text-lg text-blue-200 ml-3">Grupos</h2>
+      <ul className="space-y-2 flex flex-col mt-[20px] pr-4">
+        <li className="text-blue-300 flex items-center hover:bg-blue-800/20 px-2 py-3 rounded-md transition">
+          <div className="w-[35px] h-[35px] bg-blue-500 rounded-sm inline-block mr-2"></div>
+          <div>
+            <p className="text-sm font-bold">Nome do Grupo</p>
+            <p className="text-xs text-blue-400">Descrição do grupo</p>
+          </div>
+        </li>
+        <li className="text-blue-300 flex items-center hover:bg-blue-800/20 px-2 py-3 rounded-md transition">
+          <div className="w-[35px] h-[35px] bg-blue-500 rounded-sm inline-block mr-2"></div>
+          <div>
+            <p className="text-sm font-bold">Nome do Grupo</p>
+            <p className="text-xs text-blue-400">Descrição do grupo</p>
+          </div>
+        </li>
+        <li className="text-blue-300 flex items-center hover:bg-blue-800/20 px-2 py-3 rounded-md transition">
+          <div className="w-[35px] h-[35px] bg-blue-500 rounded-sm inline-block mr-2"></div>
+          <div>
+            <p className="text-sm font-bold">Nome do Grupo</p>
+            <p className="text-xs text-blue-400">Descrição do grupo</p>
+          </div>
+        </li>
+        <li className="text-blue-300 flex items-center hover:bg-blue-800/20 px-2 py-3 rounded-md transition">
+          <div className="w-[35px] h-[35px] bg-blue-500 rounded-sm inline-block mr-2"></div>
+          <div>
+            <p className="text-sm font-bold">Nome do Grupo</p>
+            <p className="text-xs text-blue-400">Descrição do grupo</p>
+          </div>
+        </li>
+        <li className="text-blue-300 flex items-center hover:bg-blue-800/20 px-2 py-3 rounded-md transition">
+          <div className="w-[35px] h-[35px] bg-blue-500 rounded-sm inline-block mr-2"></div>
+          <div>
+            <p className="text-sm font-bold">Nome do Grupo</p>
+            <p className="text-xs text-blue-400">Descrição do grupo</p>
+          </div>
+        </li>
+        <li className="text-blue-300 flex items-center hover:bg-blue-800/20 px-2 py-3 rounded-md transition">
+          <div className="w-[35px] h-[35px] bg-blue-500 rounded-sm inline-block mr-2"></div>
+          <div>
+            <p className="text-sm font-bold">Nome do Grupo</p>
+            <p className="text-xs text-blue-400">Descrição do grupo</p>
+          </div>
+        </li>
+        <li className="text-blue-300 flex items-center hover:bg-blue-800/20 px-2 py-3 rounded-md transition">
+          <div className="w-[35px] h-[35px] bg-blue-500 rounded-sm inline-block mr-2"></div>
+          <div>
+            <p className="text-sm font-bold">Nome do Grupo</p>
+            <p className="text-xs text-blue-400">Descrição do grupo</p>
+          </div>
+        </li>
+      </ul>
+    </div>
+  );
+}
